[ticket/16016] Add a base class for native password hashing
PHPBB3-16016
